Continue Branch Code
Continued code to see if $t0 register is equal to new line or null, if true, jump to the isEmpty branch to output the message for an empty string input. I then began the branch to check for first few characters in string. Using the load byte instruction, I again load the address in $t2 (the inputed string) to the $t0 register and use the pointer increment to begin code for a loop through.