Restart and Continue Branches
Create a restart code to be used in previous loops, which will restart both the pointer and the counter. Restarted the pointer by subtracting values in the $t2 and $t1 registers and storing it back in $t2, restarted the counter by loading 0 directly into the register. Began code for the continuation, load the value at $t2 to the $t0 register. Created pointer.