Condtions
Added other conditions for determining string length, if the value at $t0 is equal to a new line, jump to the doLoop branch which will help keep track of the counters, if the value is equal to null, do the same(jump to next branch), if the value is equal to a space, do the same, if the length of the input is equal to five, it is too long and invalid, if this is true the isTooLong branch must be entered to display the message for an input that is too long. Set to five because input can not be more than 4.